Create Steep Array
Given a list of integers. The steepness score of an index i is computed in the following way-

Steepness Score(i) =  (maximum element to the right of index i (including i) )- A[i]
Find out the total steepness score of the array, i.e - sum of steepness scores of each index in the array.

Constraints
1<= length of array<= 100000
Input
First line of the input contains N, the size of the array. the second line contains the N space separated integers representing the N elements of the array.

Output
print the total steepness score of the array.