Create Reduce_array.py
Reduce Array
You are given a list of integers. Your task is to reduce the list to a single integer. Following are the rules to do this reduction:

[ADD] Sum first two numbers and result is m
[SUB] Say third number is e3, subtract e3 from this: m-e3
[ADD] Say fourth number is e4, add to previous result: m-e3+e4
[SUB] Subtract next number from the previous result
[ADD] Add next number to previous result and so on
until you reach the end of the list
the result at this point is the output of the problem
Input
First line is n, 1 <= n <= 100, denoting the number of elements in the list

This is followed by n lines, one integer on each line

Output
One integer indicating the output