Create Parity_Minimum.py
Description
Given an array of length n which contains only positive integers, let S be the sum of the digits of the minimal element of the array.
Return 0 if S is odd, otherwise return 1.

Input format
First line contains a positive integer n, which denotes the number of elements in the array. This is followed by n lines each containing one integer denoting one element of the array.

Output format
One line which is 0 or 1.