Create FoldArray.py
You are given an array of size n, 0 <= n <= 100. Imagine the array as a length of rope. You have to fold the rope at the middle. You are also given an input integer num_folds that specifies the number of times you should fold your array.

To illustrate more, say the given array is 1,4,9. Folding it in the middle results in 10, 4 as 9 and 1 get combined by the fold.

Say the given array is 1,10,20,21,2. Folding it would lead to: 3, 31, 20 as 2, 1 have combined and 10, 21 have combined.

Based on the num_folds repeat folding.

Input.
The first line contains n, the number of elements in the array. 0 <= n <= 200

This is followed by n lines, each containing one integer

The last line contains m, the number of folds to do. 0 <= m <= 6

Output
The first line contains k, the number of elements in the final array

This is followed by k lines, each containing one integer element of the output array