Create MAX_Product.py
You are given two array A and B. You have to select two integers, one from A and one from B. Let them me A[i] (0<=i< length of array A) and B[j] (0<=j< length of array B). you have to find the maximum value of |A[i]XB[j]|.

note:- |x| is absolute value of x.

Input
first line contains the size of the array A which is followed by size of array B in line two. Next two lines contain the two arrays A,B.

1<=length of A,B<=10^5

-10^9< A[i],B[j]< 10^9

Output
Print the maximum value of |A[i]XB[j]|