Create ColorSort.py
Given an array nums with n objects colored red, white, or blue, sort them in-place so that objects of the same color are adjacent, with the colors in the order red, white, and blue.

Here, we will use the integers 0, 1, and 2 to represent the color red, white, and blue respectively.

Follow up:

Solve this problem without using the library's sort function
Come up with a one-pass algorithm using only O(1) constant space
Input
First line denotes n, the size of the array nums. The next n lines contains elements of the array nums.

Output
n lines denoting the elements of the resultant array.

Example
Input:

6
2
0
2
1
1
0
Output:

0
0
1
1
2
2
Constraints:
n == nums.length
1 <= n <= 1000000
nums[i] is 0, 1, or 2.