Create IncreasingSubbArray.py
Given an array containing n numbers. The problem is to find the length of the longest contiguous subarray such that every element in the subarray is strictly greater than its previous element in the same subarray and also print the resulting starting and ending index. Array is 1 index based.

Input
Fisrt line has N i.e number of elements then follow N numbers in a line.(N >= 2)

Output
Print a list of 3 elements where the first one is the length of the subarray and the next 2 are the starting and the ending index Note : all elements of the list must be printed in the same line.

Example
Input:

3

2 3 1

Output:

2 1 2

The longest Increasing subarray is [2, 3] and the starting index is[1 , 2]