Create GCD:Recursive.py
Description
Write a program that computes the GCD of two positive integers using Euclid's algorithm. Your implementation should be recursive in nature. Given two positive integers a, b (a >= b), Euclid's algorithm is defined as follows.

GCD(a, b) = b, if b is a divisor of a
          = GCD(b, a%b), otherwise
Input/Output format
You only need to implement the gcd function provided in the template. The template takes care of reading the input, passing it on to gcd function and printing the result.