Create Palindrome:Recursive.py
Description
A palindrome is a string that reads the same forward and backwards.
Write a function that reads a string and returns

True if the given string is a palindrome
False if the given string is not a palindrome
IMPORTANT
Your function should NOT use the following.

Loops (you can use a loop to read the test cases)
In-built functions like reverse
Input format
First line contains n, the number of test cases.
This is followed by n lines , each containing one input string.

Output format
n lines, each containing either True or False, depending on whether the corresponding input string is a palindrome or not

Sample input
2
tenet
ayashdashd
Sample output
True
False
Explanation
First line contains 2, meaning there are 2 test cases.
tenet is a palindrome. So first line of output is True.
ayashdashd is not a palindrome. So second line of output is False.