Create BalancedParentheses.py
Description
Write a program to generate all the distinct valid combinations that are possible using n pairs of parentheses.

Input/Output format
You only need to implement the getAllValidCombinations function provided in the template. The template takes care of reading the input, passing it on to getAllValidCombinations function and printing the result.