Create Neighbours.py
There are two groups of people, A and B. There is a lack of peace between the two, so to reduce this tension, a panel was made. A panel is basically a seating arrangement in which people from either A or B category can sit. However, the restriction imposed to increase communication between the two groups limit the people in the way they can sit.

At most 2 people from same group can sit adjacent to each other.
At a time any number of people from any group can be present in the panel.
Now you have been given a panel of size 'N'. You have to figure out all the valid seating arrangements. Lets say a person from 'A' is represented by 'a' and a person from 'B' is represented by 'b'. Print all such valid seating arrangements in lexicographically increasing order.(here it is same as sorted order)

ex- aabaabba is a valid seating arrangement as not more than 2 people from any group are adjacent. but aabbbab is not a valid seating arrangement as 3 b's are sitting adjacently.

Constraints
1<= N<= 30

Input
input contains 1 line which contains an integer N representing te size of the panel.

Output
Print all the possible valid seating arrangements possible(in form of a string of a's and b's) in lexicographically increasing order.

Example
Input:

4

Output:

aaba

aabb

abaa

abab

abba

baab

baba

babb

bbaa

bbab