Create SearchinaLL.py
Description
You are given a linked list and a particular key to search for. Return True if it exists or False otherwise.

Each node of the linked list has a key field and a next field

Do not worry about reading the input in this problem. Simply fill the function search with your implementation.

Do not worry about printing output. Simply make sure that the function search returns the correct Boolean.

Explanation
Say the linked list is 76 -> 66 -> 11 -> 73 -> 77 and the given key is 73, output is True