Create DetectLoop.py
Given a head pointer to the linked list, check if the linked list has loop or not. If it has loop, function should return 1, else it should return 0.

Note:
This is a functional problem, so you don't need to worry about the input and output format. Simply take care of the detectLoop function.

Example & Explanation:
Suppose, given Linked List is:

76 -> 66 -> 11 -> 73 -> 77 
       ^                 |
       |_________________|  
Then, the output should be:

1
Because, the given linked list consists of a loop!