Create LoopLength.py
Given a head pointer to the linked list, complete the function detectAndCountLoop() that checks whether a given Linked List contains loop and if loop is present then returns count of nodes in loop. For example, the loop is present in below-linked list and length of the loop is 4. If the loop is not present, then the function should return 0.

Note:
This is a functional problem, so you don't need to worry about the input and output format. Simply take care of the detectAndCountLoop function.

Example & Explanation:
Suppose, given Linked List is:

76 -> 66 -> 11 -> 73 -> 77 
       ^                 |
       |_________________|  
Then, the output should be:

4
Because, the given linked list consists of a loop and the length of the loop is 4!